fixed ai conversion warnings

+++ Extinguish/Extinguish/AI.cpp
@@ -499,7 +499,7 @@ void AI::GetBall()
 			myNode = aiPath->FindClosest(me->GetTransform()->GetWorldPosition());
 			tarNode = aiPath->FindBallNode(ball->GetTransform()->GetWorldPosition());
 			path = aiPath->CreatePath(myNode, tarNode);
-			pathIndex = path.size() - 2;
+			pathIndex = (int)(path.size() - 2);
 			validPath = true;
 		}
 
@@ -658,7 +658,7 @@ bool AI::RunTo(GameObject *target)
 				else tarNode = aiPath->FindClosest(target->GetTransform()->GetWorldPosition());
 
 				path = aiPath->CreatePath(myNode, tarNode);
-				pathIndex = path.size() - 2;
+				pathIndex = (int)(path.size() - 2);
 				validPath = true;
 			}
